Extract a reusable boilerplate file from scrolling.rs
